Revert "move vyper pytest classes to separate file"
This reverts commit 4578ffd55adbca23199aab3cdc25763e6b438592.
