+++ tests/conftest.py
@@ -11,11 +11,12 @@ from web3 import Web3
 from web3.providers.eth_tester import EthereumTesterProvider
 
 from vyper import compiler
+from vyper.cli.vyper_compile import compile_files
 from vyper.codegen.lll_node import LLLnode
+from vyper.exceptions import VyperException
 from vyper.lll import compile_lll, optimizer
 
 from .base_conftest import VyperContract, _get_contract, zero_gas_price_strategy
-from .vyper_pytest import VyperFile
 
 # Import the base_conftest fixtures
 pytest_plugins = ["tests.base_conftest", "tests.fixtures.memorymock"]
@@ -240,3 +241,48 @@ def vyper_contract(request):
 def pytest_collect_file(parent, path):
     if path.ext == ".vy" and path.basename.startswith("test"):
         return VyperFile.from_parent(parent, fspath=path)
+
+
+class VyperFile(pytest.File):
+    def collect(self):
+        filepath = str(self.fspath)
+        yield VyperTestItem.from_parent(self, name=filepath)
+
+
+class VyperTestItem(pytest.Item):
+    def __init__(self, name, parent):
+        super().__init__(name, parent)
+        self.test_condition = self.name[:-3].split("_")[-1]
+
+    def runtest(self):
+        try:
+
+            compile_files(
+                [self.fspath],
+                ("bytecode",),
+            )
+            self.test_result = "Success"
+
+        except VyperException as v:
+            self.test_result = v.__class__.__name__
+
+        except Exception:
+            self.test_result = "Fail"
+
+        if self.test_condition != self.test_result:
+            raise VyperTestException(self, self.name)
+
+    def repr_failure(self, excinfo):
+        if isinstance(excinfo.value, VyperTestException):
+            return (
+                f"Test failed : {self.name}\n"
+                f"Expected: {self.test_condition}\n"
+                f"Actual: {self.test_result}\n"
+            )
+
+    def reportinfo(self):
+        return self.fspath, 0, f"Vyper test file: {self.name}"
+
+
+class VyperTestException(Exception):
+    pass
